+++ routes/api.php
@@ -8,9 +8,9 @@ Route::post('/auth/login', NameAuthController::class)->name('api.auth.login');
 Route::post('/auth/register', [PlayerController::class, 'store'])->name('api.auth.register');
 Route::post('/player', [PlayerController::class, 'store'])->name('api.player.store');
 Route::get('/player/name/{name}', [PlayerController::class, 'findByName'])->name('api.player.find-by-name');
+Route::get('/player/{name}/report', [PlayerController::class, 'report'])->name('api.player.report');
+Route::get('/player/{name}/certificate', [PlayerController::class, 'certificate'])->name('api.player.certificate');
 
 Route::middleware('auth:sanctum')->group(function () {
     Route::put('/player/{name}/finish', [PlayerController::class, 'finish'])->name('api.player.finish');
-    Route::get('/player/{name}/report', [PlayerController::class, 'report'])->name('api.player.report');
-    Route::get('/player/{name}/certificate', [PlayerController::class, 'certificate'])->name('api.player.certificate');
 });